feat/ui: change components paths
